requirements.txt
aiofiles==0.4.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Deprecated==1.2.5
h11==0.8.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
PyGithub==1.43.7
PyJWT==1.7.1
python-decouple==3.1
python-multipart==0.0.5
PyYAML==5.1
requests==2.21.0
six==1.12.0
starlette==0.12.0
urllib3==1.24.3
uvicorn==0.7.1
websockets==7.0
wincertstore==0.2
wrapt==1.11.1
